move mcp tools into agent package
