+++ src/pkg/agent/tools/destroy.go
