+++ src/pkg/agent/tools/destroy_test.go
