+++ src/pkg/agent/tools/estimate.go
