+++ src/pkg/agent/tools/estimate_test.go
