+++ src/pkg/agent/tools/interfaces.go
