+++ src/pkg/agent/tools/login_test.go
