+++ src/pkg/agent/tools/logs.go
