+++ src/pkg/agent/tools/removeConfig.go
