+++ src/pkg/agent/tools/removeConfig_test.go
