+++ src/pkg/agent/tools/services_test.go
