+++ src/pkg/mcp/tests/client_test.go
@@ -14,9 +14,9 @@ import (
 	m3mcp "github.com/mark3labs/mcp-go/mcp"
 	"google.golang.org/protobuf/types/known/emptypb"
 
+	"github.com/DefangLabs/defang/src/pkg/agent/tools"
 	cliClient "github.com/DefangLabs/defang/src/pkg/cli/client"
 	"github.com/DefangLabs/defang/src/pkg/mcp"
-	"github.com/DefangLabs/defang/src/pkg/mcp/tools"
 	typepb "github.com/DefangLabs/defang/src/protos/google/type"
 	defangv1 "github.com/DefangLabs/defang/src/protos/io/defang/v1"
 	"github.com/DefangLabs/defang/src/protos/io/defang/v1/defangv1connect"
